+++ packages/stream_chat_flutter/example/lib/split_view.dart
@@ -103,7 +103,7 @@ class ChannelListPage extends StatelessWidget {
                 : null,
             filter: Filter.in_(
               'members',
-              [StreamChat.of(context).user!.id],
+              [StreamChat.of(context).currentUser!.id],
             ),
             sort: const [SortOption('last_message_at')],
             pagination: const PaginationParams(