+++ packages/stream_chat_flutter/test/src/channel_header_test.dart
@@ -18,8 +18,9 @@ void main() {
       final lastMessageAt = DateTime.parse('2020-06-22 12:00:00');
 
       when(() => client.state).thenReturn(clientState);
-      when(() => clientState.user).thenReturn(user);
-      when(() => clientState.userStream).thenAnswer((_) => Stream.value(user));
+      when(() => clientState.currentUser).thenReturn(user);
+      when(() => clientState.currentUserStream)
+          .thenAnswer((_) => Stream.value(user));
       when(() => channel.lastMessageAt).thenReturn(lastMessageAt);
       when(() => channel.state).thenReturn(channelState);
       when(() => channel.client).thenReturn(client);
@@ -82,8 +83,9 @@ void main() {
       final lastMessageAt = DateTime.parse('2020-06-22 12:00:00');
 
       when(() => client.state).thenReturn(clientState);
-      when(() => clientState.user).thenReturn(user);
-      when(() => clientState.userStream).thenAnswer((_) => Stream.value(user));
+      when(() => clientState.currentUser).thenReturn(user);
+      when(() => clientState.currentUserStream)
+          .thenAnswer((_) => Stream.value(user));
       when(() => channel.lastMessageAt).thenReturn(lastMessageAt);
       when(() => channel.state).thenReturn(channelState);
       when(() => channel.client).thenReturn(client);
@@ -149,8 +151,9 @@ void main() {
       final lastMessageAt = DateTime.parse('2020-06-22 12:00:00');
 
       when(() => client.state).thenReturn(clientState);
-      when(() => clientState.user).thenReturn(user);
-      when(() => clientState.userStream).thenAnswer((_) => Stream.value(user));
+      when(() => clientState.currentUser).thenReturn(user);
+      when(() => clientState.currentUserStream)
+          .thenAnswer((_) => Stream.value(user));
       when(() => channel.lastMessageAt).thenReturn(lastMessageAt);
       when(() => channel.state).thenReturn(channelState);
       when(() => channel.client).thenReturn(client);
@@ -217,8 +220,9 @@ void main() {
       final lastMessageAt = DateTime.parse('2020-06-22 12:00:00');
 
       when(() => client.state).thenReturn(clientState);
-      when(() => clientState.user).thenReturn(user);
-      when(() => clientState.userStream).thenAnswer((_) => Stream.value(user));
+      when(() => clientState.currentUser).thenReturn(user);
+      when(() => clientState.currentUserStream)
+          .thenAnswer((_) => Stream.value(user));
       when(() => channel.lastMessageAt).thenReturn(lastMessageAt);
       when(() => channel.state).thenReturn(channelState);
       when(() => channel.client).thenReturn(client);
@@ -293,8 +297,9 @@ void main() {
       final lastMessageAt = DateTime.parse('2020-06-22 12:00:00');
 
       when(() => client.state).thenReturn(clientState);
-      when(() => clientState.user).thenReturn(user);
-      when(() => clientState.userStream).thenAnswer((_) => Stream.value(user));
+      when(() => clientState.currentUser).thenReturn(user);
+      when(() => clientState.currentUserStream)
+          .thenAnswer((_) => Stream.value(user));
       when(() => channel.lastMessageAt).thenReturn(lastMessageAt);
       when(() => channel.state).thenReturn(channelState);
       when(() => channel.client).thenReturn(client);
@@ -360,8 +365,9 @@ void main() {
       final lastMessageAt = DateTime.parse('2020-06-22 12:00:00');
 
       when(() => client.state).thenReturn(clientState);
-      when(() => clientState.user).thenReturn(user);
-      when(() => clientState.userStream).thenAnswer((_) => Stream.value(user));
+      when(() => clientState.currentUser).thenReturn(user);
+      when(() => clientState.currentUserStream)
+          .thenAnswer((_) => Stream.value(user));
       when(() => channel.lastMessageAt).thenReturn(lastMessageAt);
       when(() => channel.state).thenReturn(channelState);
       when(() => channel.client).thenReturn(client);