+++ packages/stream_chat_flutter/test/src/gradient_avatar_test.dart
@@ -15,7 +15,7 @@ void main() {
       final clientState = MockClientState();
 
       when(() => client.state).thenReturn(clientState);
-      when(() => clientState.user).thenReturn(OwnUser(id: 'user-id'));
+      when(() => clientState.currentUser).thenReturn(OwnUser(id: 'user-id'));
 
       await tester.pumpWidget(
         MaterialApp(