+++ packages/stream_chat_flutter_core/lib/src/stream_chat_core.dart
@@ -98,10 +98,20 @@ class StreamChatCoreState extends State<StreamChatCore>
   Widget build(BuildContext context) => widget.child;
 
   /// The current user
-  User? get user => client.state.user;
+  @Deprecated('Use `.currentUser` instead, Will be removed in future releases')
+  User? get user => client.state.currentUser;
 
   /// The current user as a stream
-  Stream<User?> get userStream => client.state.userStream;
+  @Deprecated(
+    'Use `.currentUserStream` instead, Will be removed in future releases',
+  )
+  Stream<User?> get userStream => client.state.currentUserStream;
+
+  /// The current user
+  User? get currentUser => client.state.currentUser;
+
+  /// The current user as a stream
+  Stream<User?> get currentUserStream => client.state.currentUserStream;
 
   StreamSubscription<ConnectivityResult>? _connectivitySubscription;
 
@@ -126,7 +136,7 @@ class StreamChatCoreState extends State<StreamChatCore>
         if (!_isInForeground) return;
         if (_isConnectionAvailable) {
           if (client.wsConnectionStatus == ConnectionStatus.disconnected &&
-              user != null) {
+              currentUser != null) {
             client.openConnection();
           }
         } else {
@@ -163,7 +173,7 @@ class StreamChatCoreState extends State<StreamChatCore>
       AppLifecycleState.resumed,
       AppLifecycleState.inactive,
     ].contains(state);
-    if (user != null) {
+    if (currentUser != null) {
       if (_isInForeground) {
         _onForeground();
       } else {